fix: model subject

+++ student-management-system-server/src/main/java/com/example/projectbase/domain/dto/response/SubjectResponseDTO.java
@@ -11,9 +11,9 @@ public class SubjectResponseDTO {
     private String id; // ID của Subject
     private String name; // Tên của Subject
     private String code; // Mã môn học
-    private int numberOfCredits; // Số tín chỉ
-    private int regularCoefficient; // Hệ số điểm chuyên cần
-    private int midTermCoefficient; // Hệ số điểm giữa kỳ
-    private int finalCoefficient; // Hệ số điểm cuối kỳ
+    private double numberOfCredits; // Số tín chỉ
+    private double regularCoefficient; // Hệ số điểm chuyên cần
+    private double midTermCoefficient; // Hệ số điểm giữa kỳ
+    private double finalCoefficient; // Hệ số điểm cuối kỳ
     private String description; // Mô tả môn học
 }
